move template functions to a single file
